+++ modules/app/components/__tests__/AccountSelect.spec.tsx
@@ -3,19 +3,9 @@ import userEvent from '@testing-library/user-event';
 import { screen } from '@testing-library/react';
 import WrappedAccountSelect from 'modules/app/components/layout/header/AccountSelect';
 import { useRouter } from 'next/router';
-import { ConnectionType } from 'modules/web3/connections';
 
 jest.mock('next/router');
 
-const walletName = 'MetaMask';
-jest.mock('modules/web3/constants/wallets', () => ({
-  SUPPORTED_WALLETS: {
-    [walletName]: {
-      connectionType: 'INJECTED',
-      name: walletName
-    }
-  }
-}));
 jest.mock('modules/web3/connections', () => ({ connectorToWalletName: () => null }));
 
 describe('Account select', () => {
@@ -29,7 +19,7 @@ describe('Account select', () => {
 
     userEvent.click(connectButton);
 
-    const metamaskButton = await screen.findByText(walletName);
+    const metamaskButton = await screen.findByText('MetaMask');
     expect(metamaskButton).toBeInTheDocument();
   });
 });
