+++ modules/app/components/layout/header/AccountSelect.tsx
@@ -19,8 +19,8 @@ import { useRouter } from 'next/router';
 import { InternalLink } from 'modules/app/components/InternalLink';
 import { isAndroid, isIOS } from 'react-device-detect';
 import { getExecutiveVotingWeightCopy } from 'modules/polling/helpers/getExecutiveVotingWeightCopy';
-import { SUPPORTED_WALLETS, WalletName } from 'modules/web3/constants/wallets';
-import { connectorToWalletName, getConnection, ConnectionType } from 'modules/web3/connections';
+import { SUPPORTED_WALLETS, WalletName, ConnectionType } from 'modules/web3/constants/wallets';
+import { connectorToWalletName, getConnection } from 'modules/web3/connections';
 import { AnalyticsContext } from 'modules/app/client/analytics/AnalyticsContext';
 import { isSupportedChain } from 'modules/web3/helpers/chain';
 import logger from 'lib/logger';
