+++ modules/web3/components/Web3Provider.tsx
@@ -1,6 +1,6 @@
 import { Web3ReactHooks, Web3ReactProvider } from '@web3-react/core';
 import { Connector } from '@web3-react/types';
-import { Connection } from 'modules/web3/connections';
+import { Connection } from 'modules/web3/types/connection';
 import { useEagerlyConnect } from 'modules/web3/hooks/useEagerlyConnect';
 import { useOrderedConnections } from 'modules/web3/hooks/useOrderedConnections';
 import { ReactNode, useMemo } from 'react';
