+++ modules/web3/types/wallets.d.ts
@@ -1,8 +1,6 @@
-import { ConnectionType } from 'modules/web3/connections';
-import { WalletName } from 'modules/web3/constants/wallets';
+import { WalletName, ConnectionType } from 'modules/web3/constants/wallets';
 
 export interface WalletInfo {
-  // connection: Connection;
   connectionType: ConnectionType;
   name: WalletName;
   deeplinkUri?: string;
